+++ src/main/kotlin/com/christopherosthues/starwarsprogressbar/configuration/components/UiOptionsPanel.kt
@@ -3,6 +3,7 @@ package com.christopherosthues.starwarsprogressbar.configuration.components
 import com.christopherosthues.starwarsprogressbar.StarWarsBundle
 import com.christopherosthues.starwarsprogressbar.configuration.StarWarsState
 import com.christopherosthues.starwarsprogressbar.constants.BundleConstants
+import com.christopherosthues.starwarsprogressbar.constants.DEFAULT_CHANGE_VEHICLE_AFTER_PASS
 import com.christopherosthues.starwarsprogressbar.constants.DEFAULT_DRAW_SILHOUETTES
 import com.christopherosthues.starwarsprogressbar.constants.DEFAULT_ENABLE_NEW_VEHICLES
 import com.christopherosthues.starwarsprogressbar.constants.DEFAULT_SAME_VEHICLE_VELOCITY
@@ -11,12 +12,21 @@ import com.christopherosthues.starwarsprogressbar.constants.DEFAULT_SHOW_TOOLTIP
 import com.christopherosthues.starwarsprogressbar.constants.DEFAULT_SHOW_VEHICLE
 import com.christopherosthues.starwarsprogressbar.constants.DEFAULT_SHOW_VEHICLE_NAMES
 import com.christopherosthues.starwarsprogressbar.constants.DEFAULT_SOLID_PROGRESS_BAR_COLOR
+import com.christopherosthues.starwarsprogressbar.constants.DEFAULT_NUMBER_OF_PASSES_UNTIL_VEHICLE_CHANGE
+import com.intellij.ui.JBIntSpinner
 import com.intellij.ui.components.JBCheckBox
+import com.intellij.ui.components.JBPanel
+import java.awt.FlowLayout
 import java.awt.GridLayout
+import java.awt.Insets
+import javax.swing.JPanel
+import javax.swing.event.ChangeEvent
 
 private const val GAP = 5
 
-private const val NUMBER_OF_ROWS = 4
+private const val NUMBER_OF_ROWS = 5
+private const val MINIMUM_NUMBER_OF_PASSES = 1
+private const val MAXIMUM_NUMBER_OF_PASSES = 20
 
 internal class UiOptionsPanel : JTitledPanel(StarWarsBundle.message(BundleConstants.UI_OPTIONS)) {
     private val showVehicleNameCheckBox =
@@ -45,6 +55,12 @@ internal class UiOptionsPanel : JTitledPanel(StarWarsBundle.message(BundleConsta
         StarWarsBundle.message(BundleConstants.DRAW_SILHOUETTES),
         DEFAULT_DRAW_SILHOUETTES
     )
+    private val changeVehicleAfterPassCheckBox = JBCheckBox(
+        StarWarsBundle.message(BundleConstants.CHANGE_VEHICLE_AFTER_PASS),
+        DEFAULT_CHANGE_VEHICLE_AFTER_PASS
+    )
+    private val numberOfPassesUntilVehicleChangeSpinner =
+        JBIntSpinner(DEFAULT_NUMBER_OF_PASSES_UNTIL_VEHICLE_CHANGE, MINIMUM_NUMBER_OF_PASSES, MAXIMUM_NUMBER_OF_PASSES)
 
     val showVehicle: Boolean
         get() = showVehicleCheckBox.isSelected
@@ -69,6 +85,12 @@ internal class UiOptionsPanel : JTitledPanel(StarWarsBundle.message(BundleConsta
 
     val drawSilhouttes: Boolean
         get() = drawSilhouttesCheckBox.isSelected
+
+    val changeVehicleAfterPass: Boolean
+        get() = changeVehicleAfterPassCheckBox.isSelected
+
+    val numberOfPassesUntilVehicleChange: Int
+        get() = numberOfPassesUntilVehicleChangeSpinner.number
 
     init {
         layout = GridLayout(NUMBER_OF_ROWS, 2, GAP, GAP)
@@ -129,6 +151,26 @@ internal class UiOptionsPanel : JTitledPanel(StarWarsBundle.message(BundleConsta
                 drawSilhouttes
             )
         }
+        changeVehicleAfterPassCheckBox.addItemListener {
+            numberOfPassesUntilVehicleChangeSpinner.isEnabled = changeVehicleAfterPass
+            firePropertyChange(
+                this::changeVehicleAfterPass.name,
+                !changeVehicleAfterPass,
+                changeVehicleAfterPass
+            )
+        }
+        numberOfPassesUntilVehicleChangeSpinner.addChangeListener {
+            firePropertyChange(
+                this::numberOfPassesUntilVehicleChange.name,
+                numberOfPassesUntilVehicleChange,
+                numberOfPassesUntilVehicleChange
+            )
+        }
+        numberOfPassesUntilVehicleChangeSpinner.isEnabled = changeVehicleAfterPass
+
+        val passesPanel = JPanel(FlowLayout(FlowLayout.LEFT, 0, 0))
+        passesPanel.add(changeVehicleAfterPassCheckBox)
+        passesPanel.add(numberOfPassesUntilVehicleChangeSpinner)
 
         add(showVehicleCheckBox)
         add(sameVehicleVelocityCheckBox)
@@ -138,6 +180,7 @@ internal class UiOptionsPanel : JTitledPanel(StarWarsBundle.message(BundleConsta
         add(solidProgressBarColorCheckBox)
         add(showFactionCrestsCheckBox)
         add(drawSilhouttesCheckBox)
+        add(passesPanel)
     }
 
     fun updateUI(starWarsState: StarWarsState) {
@@ -149,5 +192,7 @@ internal class UiOptionsPanel : JTitledPanel(StarWarsBundle.message(BundleConsta
         enableNewVehiclesCheckBox.isSelected = starWarsState.enableNewVehicles
         solidProgressBarColorCheckBox.isSelected = starWarsState.solidProgressBarColor
         drawSilhouttesCheckBox.isSelected = starWarsState.drawSilhouettes
+        changeVehicleAfterPassCheckBox.isSelected = starWarsState.changeVehicleAfterPass
+        numberOfPassesUntilVehicleChangeSpinner.value = starWarsState.numberOfPassesUntilVehicleChange
     }
 }