+++ src/main/kotlin/com/christopherosthues/starwarsprogressbar/constants/ConfigurationConstants.kt
@@ -8,3 +8,5 @@ internal const val DEFAULT_SAME_VEHICLE_VELOCITY = false
 internal const val DEFAULT_ENABLE_NEW_VEHICLES = true
 internal const val DEFAULT_SOLID_PROGRESS_BAR_COLOR = false
 internal const val DEFAULT_DRAW_SILHOUETTES = false
+internal const val DEFAULT_CHANGE_VEHICLE_AFTER_PASS = false
+internal const val DEFAULT_NUMBER_OF_PASSES_UNTIL_VEHICLE_CHANGE = 2