+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Option to draw solid progress bar
 - Option to hide/display vehicle in progress bar
 - Option to draw only the silhouette of a vehicle
+- Option to change vehicle automatically after a predefined number of passes
 - Eta-class Shuttle to Jedi faction
 - CSS-1 Corellian Star Shuttle and Oevvaor Jet Catamaran to Galactic Republic
 - Count Dooku's Flitknot Speeder Bike, Droch-class boarding ship, Flitknot Speeder Bike and Octuptarra tri-droid to CIS